Animation Script Modularization & Extension (WIP)

+++ src/battle/battle.h
@@ -14,9 +14,9 @@ typedef struct {
 void  srLoadBattleFormation(i32 formationIndex, i32(*modelAppearCallback)(void));
 void srExecuteAIScript(i32 actorIndex, i32 scriptID, i32 unkInt);
 i32 srExecuteFormationScripts();
-i32  enqueueScriptAction(i16 characterID, i16 commandIndex, i16 relAttackIndex);
+i32  enqueueScriptAction(u8 characterID, u8 commandIndex, u16 relAttackIndex);
 
-void dispatchAutoActions(i32 partyIndex, i32 actionType);
+void dispatchAutoActions(u8 partyIndex, i32 actionType);
 u16 getActionToDispatch(const SrAutoAction& action);
 u16 setTargetMask(u8 partyIndex, const SrAutoAction& action);
